Move Authentication type to protocol
