+++ packages/liveblocks-core/src/protocol/Authentication.ts
@@ -1,0 +1,14 @@
+export type Authentication =
+  | {
+      type: "public";
+      publicApiKey: string;
+      url: string;
+    }
+  | {
+      type: "private";
+      url: string;
+    }
+  | {
+      type: "custom";
+      callback: (room: string) => Promise<{ token: string }>;
+    };
